See commit message for fixes, changes, and new features. Thanks!
Fixes
- Fixed bug in the analysis version of BFS enumeration algorithms where a variable value was being overwritten
- Fixed bug in INSTALL_PMA_project.m and README.md (small filename issue)
- Fixed bug in PMA_SortAsPerfectMatching.m where the data type was too small (int8 -> int16). However, some previous (old) examples run out of memory

Changes
- Split up PMA_ConnectedSubcatalogFilters.m into PMA_SimpleGraphFilter.m and PMA_TreeFilter.m because the subcatalog checks can be applied to different cases (connected vs. nonconnected simple graphs)
- Changed elements of the plot style in PMA_PlotTreeEnumerate.m
- Removed triangles graph invariant because it was slow and not very helpful (see label-shifted spectrum below)

Features
- Added PMA_SubcatalogEnumerationAlg_v2.m which allows for linear penalty and linear satisfaction constraints to be specified (see PMA_TEST_SubcatalogConstraints.m for a test function)
- Added label-shifted spectrum graph invariant check to PMA_RemoveIsoLabeledGraphs.m. In most cases, this filters out all unnecessary graph isomorphism checks and substantially reduces overall computation time
- Added in code to help with counting the number of generated and feasible graphs in PMA_GenerateFeasibleGraphs.m
- Added an optional subcatalog check for the existence of a bipartite graph in PMA_BipartiteSubcatalogFilters.m
- Added PMA_TreeAnalysis.m to help with the standardization of the analysis version of the enumeration algorithms
- Added OEIS examples A001190 and A008277

+++ src/pm/PMA_SortAsPerfectMatching.m
@@ -11,7 +11,7 @@
 function G = PMA_SortAsPerfectMatching(G)
 
     % first check that is this a perfect matching
-    dG = diff(int8(G),1,2);
+    dG = diff(int16(G),1,2);
     if all(dG(:,1:2:end)>0,'all')
         G = fliplr(G); % flip so even and odd vertices switch
     elseif all(dG(:,1:2:end)<0,'all')
